Corrections
Fixed SpeedTimer, so that Oil Spill actually works. Now _speed used instead of Speed.
Corrected Conveyor Belt Scripts.
Now Left Conveyor Belt no longer has complete Right conveyor belt code inside of it. Also changed transform.right to Vector3.right, as transform.right uses rotation, wheresas Vector3.right does not. May have some consideration to changing current conveyor belts into Launchers, as there has not been any way to slow down movement involved so far.
