+++ MicrowaveGame2D/Assets/ConveyorBeltRight.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2d4058d2912019c48b159def3028b980
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
