+++ MicrowaveGame2D/Assets/Scenes/EnvironmentalHazardsTestScene.unity
@@ -331,6 +331,7 @@ GameObject:
   - component: {fileID: 1217788015}
   - component: {fileID: 1217788014}
   - component: {fileID: 1217788013}
+  - component: {fileID: 1217788012}
   m_Layer: 0
   m_Name: Conveyor Belt(Left)
   m_TagString: Untagged
@@ -338,6 +339,19 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
+--- !u!114 &1217788012
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1217788011}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 77be19ad8f7a9f34281240e8de205ee9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  toLeft: {x: 0, y: 0}
 --- !u!61 &1217788013
 BoxCollider2D:
   m_ObjectHideFlags: 0
@@ -423,8 +437,8 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1217788011}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -2.95, y: -2.94, z: 0}
-  m_LocalScale: {x: 4, y: 1, z: 1}
+  m_LocalPosition: {x: -1, y: -3, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 5
@@ -440,7 +454,7 @@ GameObject:
   - component: {fileID: 1872975192}
   - component: {fileID: 1872975191}
   - component: {fileID: 1872975195}
-  - component: {fileID: 1872975194}
+  - component: {fileID: 1872975193}
   m_Layer: 0
   m_Name: Conveyor Belt(Right)
   m_TagString: Untagged
@@ -507,13 +521,13 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1872975190}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 2, y: -3, z: 0}
-  m_LocalScale: {x: 4, y: 1, z: 1}
+  m_LocalPosition: {x: 1, y: -3, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &1872975194
+--- !u!114 &1872975193
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -522,7 +536,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1872975190}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 62b10a190d26bc843bab20948cbaaea8, type: 3}
+  m_Script: {fileID: 11500000, guid: 271116ebd5730d349a102273c868b9a2, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   toRight: {x: 0, y: 0}
