+++ MicrowaveGame2D/Assets/Scripts/Hazards.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ff9e88629bf41bf4f84942da0f1779a3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
