+++ MicrowaveGame2D/Assets/Scripts/Hazards/ConveyorBeltRight.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 62b10a190d26bc843bab20948cbaaea8
+guid: 271116ebd5730d349a102273c868b9a2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
